requirements.txt
certifi
chardet
Django==1.11.29
django-braces==1.8.1
django-embed-video
djangorestframework==3.11.2
idna
olefile
Pillow==8.3.2
python3-memcached
pytz
whitenoise==3.3.1
dj-database-url
gunicorn
psycopg2-binary
python-decouple
django-redis
django-storages
boto3
django-widget-tweaks==1.4.3
google-api-python-client==1.7.4
numpy==1.19.5
pandas==1.1.5
scikit-learn==0.24.1
scipy==1.5.4
sklearn==0.0
django-webpack-loader==0.6.0
django-cors-headers==3.0.2
Markdown==3.1.1
django-taggit==1.1.0
django-taggit-serializer==0.1.7
mistune==0.8.4
pygments==2.7.4
celery==4.4.1
redis==3.4.1
django-autoslug==1.9.7
django-termsandconditions==1.2.15
django-crispy-forms==1.9.2
opencensus==0.9.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.4
opencensus-ext-django==0.7.6
azure-core==1.24.1
azure-identity==1.10.0
azure-servicebus